Add files via upload
Alice and Bob are playing a game. They have a common string SS of length NN. The players also have their individual strings AA (belonging to Alice) and BB (belonging to Bob) which are empty in the beginning. Game begins with Alice and both players take alternate turns.

In her/his turn, the player picks a single character from string SS, adds it to the end of their individual string and deletes the picked character from string SS.

The game continues until string SS is empty. Find whether there exists a sequence of moves such that the strings AA and BB are same at the end of the game